+++ app/scripts/components/search-examples.ts
@@ -4,6 +4,7 @@ import _ from "lodash"
 import statemachine from "@/statemachine"
 import { html } from "@/util"
 import settings from "@/settings"
+import { SearchExample } from "@/settings/app-settings.types"
 
 export default angular.module("korpApp").component("searchExamples", {
     template: html`
@@ -31,7 +32,7 @@ export default angular.module("korpApp").component("searchExamples", {
 
             $ctrl.$onInit = () => {
                 // Find search query examples
-                const examples = settings["frontpage"]?.["examples"]
+                const examples = settings.frontpage?.examples
                 if (examples) {
                     // Pick three random examples
                     $scope.examples = _.shuffle(examples).slice(0, 3)
@@ -41,7 +42,7 @@ export default angular.module("korpApp").component("searchExamples", {
             $ctrl.setSearch = (params: Record<string, any>) => {
                 if (params.corpus) {
                     const corpora = params.corpus.split(",")
-                    $rootScope._settings.corpusListing.select(corpora)
+                    settings.corpusListing.select(corpora)
                     $rootScope.$broadcast("corpuschooserchange", corpora)
                 }
                 if (params.cqp) {
@@ -56,11 +57,3 @@ export default angular.module("korpApp").component("searchExamples", {
 type SearchExamplesScope = IScope & {
     examples?: SearchExample[]
 }
-
-type SearchExample = {
-    label: TranslatedString | string
-    hint: TranslatedString | string
-    params: Record<string, string | number>
-}
-
-type TranslatedString = { [lang: string]: string }