+++ app/scripts/index.d.ts
@@ -1,0 +1,5 @@
+
+declare module "korp_config" {
+    const settings: import("@/settings/settings.types").Settings
+    export = settings
+}