+++ app/scripts/timeseries.ts
@@ -22,10 +22,10 @@ export function calculateYearTicks(min: number, max: number) {
 
 // Time data is fetched in data_init.js, to also provide data for search result trend diagram (?)
 /** Data size per year of all corpora. */
-export const getTimeDataPairs = (): [number, number][] => settings["time_data"][0]
+export const getTimeDataPairs = (): [number, number][] => settings.time_data[0]
 
 /** Data size of unknown year in all corpora. */
-export const getCountUndated = (): number => settings["time_data"][1]
+export const getCountUndated = (): number => settings.time_data[1]
 
 /** Get data size per year of all corpora. */
 export const getSeries = () => fromPairs(getTimeDataPairs()) as YearSeries