+++ app/scripts/util.ts
@@ -285,8 +285,8 @@ export function setDownloadLinks(xhr_settings: JQuery.AjaxSettings, result_data)
     }
     $("#download-links").append("<option value='init' rel='localize[download_kwic]'></option>")
     i = 0
-    while (i < settings["download_formats"].length) {
-        const format = settings["download_formats"][i]
+    while (i < settings.download_formats.length) {
+        const format = settings.download_formats[i]
         // NOTE: Using attribute rel="localize[...]" to localize the
         // title attribute requires a small change to
         // lib/jquery.localize.js. Without that, we could use
@@ -305,17 +305,17 @@ export function setDownloadLinks(xhr_settings: JQuery.AjaxSettings, result_data)
             query_params: xhr_settings.url,
             format,
             korp_url: window.location.href,
-            korp_server_url: settings["korp_backend_url"],
+            korp_server_url: settings.korp_backend_url,
             corpus_config: JSON.stringify(result_corpora_settings),
             corpus_config_info_keys: ["metadata", "licence", "homepage", "compiler"].join(","),
             urn_resolver: settings.urnResolver,
         }
         if ("downloadFormatParams" in settings) {
-            if ("*" in settings["download_format_params"]) {
-                $.extend(download_params, settings["download_format_params"]["*"])
+            if ("*" in settings.download_format_params) {
+                $.extend(download_params, settings.download_format_params["*"])
             }
-            if (format in settings["download_format_params"]) {
-                $.extend(download_params, settings["download_format_params"][format])
+            if (format in settings.download_format_params) {
+                $.extend(download_params, settings.download_format_params[format])
             }
         }
         option.appendTo("#download-links").data("params", download_params)
@@ -330,7 +330,7 @@ export function setDownloadLinks(xhr_settings: JQuery.AjaxSettings, result_data)
             if (!params) {
                 return
             }
-            ;($ as any).generateFile(settings["download_cgi_script"], params)
+            ;($ as any).generateFile(settings.download_cgi_script, params)
             const self = $(this)
             return setTimeout(() => self.val("init"), 1000)
         })
@@ -403,7 +403,7 @@ export function httpConfAddMethodFetch(
     url: string,
     params: Record<string, string>
 ): { url: string; request?: RequestInit } {
-    if (calcUrlLength(url, params) > settings["backendURLMaxLength"]) {
+    if (calcUrlLength(url, params) > settings.backendURLMaxLength) {
         const body = new FormData()
         for (const key in params) {
             body.append(key, params[key])